+++ src/Sample.Enquiry.Infrastructure/Data/EfRepository.cs
@@ -15,7 +15,7 @@ namespace Sample.Enquiry.Infrastructure.Data
             _dbContext = dbContext;
         }
 
-        public T GetById<T>(int id) where T : BaseEntity
+        public T GetById<T>(ulong id) where T : BaseEntity
         {
             return _dbContext.Set<T>().SingleOrDefault(e => e.Id == id);
         }
